moved mask to correct folder
